refactor the code to more readible
