+++ src/main/java/leyans/RidersHub/Controller/Api/MapboxController.java
@@ -1,4 +1,4 @@
-package leyans.RidersHub.Controller;
+package leyans.RidersHub.Controller.Api;
 
 import leyans.RidersHub.Service.MapService.MapBox.MapboxService;
 import org.springframework.beans.factory.annotation.Autowired;
