+++ src/main/java/leyans/RidersHub/Service/RideJoinRequestService.java
@@ -90,16 +90,14 @@ public class RideJoinRequestService {
     @Transactional(readOnly = true)
     public List<JoinResponseDTO> getJoinRequestsByRideId(Integer generatedRidesId, String requestingUsername) {
         try {
-            // Find the ride first
             Rides ride = riderUtil.findRideById(generatedRidesId);
 
-            // Check if the requesting user is the owner of the ride
             if (!ride.getUsername().getUsername().equals(requestingUsername)) {
                 throw new RuntimeException("Only the ride owner can view join requests");
             }
 
-            // If authorized, proceed to get join requests
             List<RideJoinRequest> requests = rideJoinRequestRepository.findByGeneratedRidesId_GeneratedRidesId(generatedRidesId);
+
             return requests.stream()
                     .map(this::convertToJoinResponseDTO)
                     .collect(Collectors.toList());
@@ -110,6 +108,7 @@ public class RideJoinRequestService {
             throw new RuntimeException("Failed to retrieve join requests", e);
         }
     }
+
 
     private JoinResponseDTO convertToJoinResponseDTO(RideJoinRequest request) {
         JoinResponseDTO responseDTO = new JoinResponseDTO();
